third_party/WebKit/LayoutTests/fast/html/imports/rel-import-to-style.html

<!DOCTYPE html>
<html>
<script src="../../../resources/js-test.js"></script>
<script src="resources/import-helpers.js"></script>
<head>
<link id="target", rel="import" href="resources/hello.html">
</head>
<body>
<script>
waitAndTest([
    {
        description: "Updating attribute value shouldn't invalidate loaded import.",
        setup: function()
        {
            debug("Before update:");
            shouldBeNull("target.sheet");
            shouldBeNonNull("target.import");
            target.setAttribute("href", "resources/hello.css");
            target.setAttribute("rel", "stylesheet");
            window.placeholder = createPlaceholder();
        },
        ready: function() { return window.placeholder.import; },
        test: function()
        {
            debug("After update:");
            shouldBeNull("target.sheet");
            shouldBeNonNull("target.import");
        }
    }
]);
</script>
</body>
</html>
